TIP-856: Move Pim\Buffer spec resources to Akeneo\Tool
